+++ airbyte-cdk/python/airbyte_cdk/destinations/vector_db_based/embedder.py
@@ -6,6 +6,7 @@ from abc import ABC, abstractmethod
 from typing import List, Optional
 
 from airbyte_cdk.destinations.vector_db_based.config import (
+    AzureOpenAIEmbeddingConfigModel,
     CohereEmbeddingConfigModel,
     FakeEmbeddingConfigModel,
     FromFieldEmbeddingConfigModel,
@@ -117,6 +118,33 @@ class FakeEmbedder(Embedder):
     @property
     def embedding_dimensions(self) -> int:
         # use same vector size as for OpenAI embeddings to keep it realistic
+        return OPEN_AI_VECTOR_SIZE
+
+
+class AzureOpenAIEmbedder(Embedder):
+    def __init__(self, config: AzureOpenAIEmbeddingConfigModel):
+        super().__init__()
+        # Client is set internally
+        self.embeddings = OpenAIEmbeddings(openai_api_key=config.openai_key, chunk_size=8191, max_retries=15, openai_api_type="azure", openai_api_version="2023-05-15", openai_api_base=config.api_base, deployment=config.deployment)  # type: ignore
+
+    def check(self) -> Optional[str]:
+        old_retries = self.embeddings.max_retries
+        try:
+            # Set retries to once to fail quickly in case the base URL is wrong as it will enter a backoff loop otherwise
+            self.embeddings.max_retries = 1
+            self.embeddings.embed_query("test")
+        except Exception as e:
+            return format_exception(e)
+        finally:
+            self.embeddings.max_retries = old_retries
+        return None
+
+    def embed_chunks(self, chunks: List[Chunk]) -> List[List[float]]:
+        return self.embeddings.embed_documents([chunk.page_content for chunk in chunks])
+
+    @property
+    def embedding_dimensions(self) -> int:
+        # vector size produced by text-embedding-ada-002 model
         return OPEN_AI_VECTOR_SIZE